Add a `#[queryable]` macro for generating trait query impls (#6)
Added the attribute macro `#[queryable]` to generate trait query impls for a trait. This replaces the `impl_trait_query!` declarative macro. This also modifies the trait definition, adding various `'static` bounds necessary for interfacing with `bevy`. As an escape hatch, this can be disabled by using the form `#[queryable(no_bounds)]`, although I can not think of any actual uses for this.