Code reorganization to facilitate planned `Added/Changed` extensions (#29)
`lib.rs` is becoming quite unwieldy with the new extensions planned to `Added/Changed`: for instance, there will be new `WorldQuery` adapters, `ReadTraits` iterators, and new `WriteTraits` iterators. In order to make the file easier to navigate, and thereby mitigate the risk of errors, I think it will be useful to reorganize the code.